Refactor code into a new `PsiElementAssociations` class
