--- src/main/kotlin/org/openrewrite/kotlin/internal/PsiInfo.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.kotlin.internal
-
-import java.util.*
-
-class PsiInfo(
-    val startOffset: Int,
-    val endOffset: Int
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is PsiInfo) return false
-        return startOffset == other.startOffset && endOffset == other.endOffset
-    }
-
-    override fun hashCode(): Int {
-        return Objects.hash(startOffset, endOffset)
-    }
-
-    override fun toString(): String {
-        return "PSI([$startOffset,$endOffset])"
-    }
-}
